Add class list pane
