+++ src/pages/school-home/SchoolHomeDesktop.vue
@@ -27,24 +27,10 @@
         <h2 class="text-h5 text-center q-mt-none q-mb-md">
           Klasy
         </h2>
-        <div class="col-shrink overflow-auto">
-          <class-list
-            :items="data.classes ?? []"
-          />
-        </div>
-        <q-space />
-        <q-btn
-          no-caps
-          :outline="!$q.dark.isActive"
-          :color="$q.dark.isActive ? 'indigo-9' : 'primary'"
-          class="full-width q-mt-md"
-          :to="combinedRoute"
-        >
-          Zestawienie klas
-        </q-btn>
-        <push-banner
-          v-if="isVLo"
-          class="q-mt-md"
+        <class-list
+          class="row-fill"
+          :items="data.classes ?? []"
+          :show-push-banner="isVLo"
         />
       </div>
       <div class="col-fill full-height column no-wrap">
@@ -79,30 +65,16 @@
 </template>
 
 <script lang="ts" setup>
-import { computed } from 'vue';
-import { useRoute } from 'vue-router';
-import { routeNames } from 'src/router/route-constants';
-import PushBanner from 'components/PushBanner.vue';
 import VLoMapView from 'components/lists/VLoMapView.vue';
 import ClassList from 'components/lists/ClassList.vue';
 import UnitList from 'components/lists/UnitList.vue';
 import type { Data } from 'layouts/SchoolLayout.vue';
-
-const route = useRoute();
 
 defineProps<{
   data: Data | null;
   isVLo: boolean;
 }>();
 
-const combinedRoute = computed(() => ({
-  name: routeNames.combinedTimetable,
-  params: route.params,
-}));
-const selectRoomRoute = computed(() => ({
-  name: routeNames.selectRoom,
-  params: route.params,
-}));
 const styleFn = (topMargin: number, height: number) => ({ height: `${height - topMargin}px` });
 </script>
